refactor: rename multiple streams stream in common
