+++ src/libecalc/common/multiple_streams_stream.py
